# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MirrorTech POS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface RefreshInput {
  refreshToken: string;
}

export interface UserProfile {
  id: string;
  username: string;
  email: string;
  role: string;
  /** @nullable */
  locationId?: string | null;
}

export interface AuthResponse {
  token: string;
  refreshToken: string;
  user: UserProfile;
}

export interface Location {
  id: string;
  name: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  isActive: boolean;
  createdAt?: string;
}

export interface LocationInput {
  name: string;
  address?: string;
  phone?: string;
}

export interface Category {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  createdAt?: string;
}

export interface CategoryInput {
  name: string;
  description?: string;
}

export interface InventoryItem {
  id: string;
  name: string;
  /** @nullable */
  sku?: string | null;
  /** @nullable */
  description?: string | null;
  price: string;
  /** @nullable */
  cost?: string | null;
  quantity: number;
  minQuantity?: number;
  /** @nullable */
  locationId?: string | null;
  /** @nullable */
  categoryId?: string | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  unit?: string | null;
  isActive?: boolean;
  createdAt?: string;
}

export interface InventoryInput {
  name: string;
  sku?: string;
  description?: string;
  price: string;
  cost?: string;
  quantity?: number;
  minQuantity?: number;
  locationId?: string;
  categoryId?: string;
  unit?: string;
}

export interface InventoryUpdate {
  name?: string;
  sku?: string;
  description?: string;
  price?: string;
  cost?: string;
  quantity?: number;
  minQuantity?: number;
  locationId?: string;
  categoryId?: string;
  unit?: string;
  isActive?: boolean;
}

export interface CartItem {
  itemId: string;
  name: string;
  quantity: number;
  price: string;
  total?: string;
}

export type TransactionInputPaymentMethod = typeof TransactionInputPaymentMethod[keyof typeof TransactionInputPaymentMethod];


export const TransactionInputPaymentMethod = {
  cash: 'cash',
  momo: 'momo',
  card: 'card',
} as const;

export interface TransactionInput {
  locationId: string;
  items: CartItem[];
  subtotal: string;
  taxAmount: string;
  total: string;
  paymentMethod: TransactionInputPaymentMethod;
  momoPhone?: string;
  momoNetwork?: string;
  momoReference?: string;
  customerName?: string;
  customerPhone?: string;
  notes?: string;
}

export interface Transaction {
  id: string;
  receiptNumber: string;
  locationId: string;
  /** @nullable */
  locationName?: string | null;
  /** @nullable */
  cashierId?: string | null;
  /** @nullable */
  cashierName?: string | null;
  items: CartItem[];
  subtotal: string;
  taxAmount: string;
  total: string;
  paymentMethod: string;
  paymentStatus: string;
  /** @nullable */
  momoPhone?: string | null;
  /** @nullable */
  momoNetwork?: string | null;
  /** @nullable */
  momoReference?: string | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerPhone?: string | null;
  /** @nullable */
  notes?: string | null;
  isVoided?: boolean;
  /** @nullable */
  voidReason?: string | null;
  createdAt: string;
}

export interface TransactionList {
  data: Transaction[];
  total: number;
}

export interface VoidInput {
  reason: string;
}

export interface Receipt {
  transaction: Transaction;
  location: Location;
  graReceiptNumber?: string;
  /** @nullable */
  qrCode?: string | null;
}

export interface SyncInput {
  transactions: TransactionInput[];
}

export interface SyncResult {
  synced: number;
  failed: number;
  errors?: string[];
}

export interface ShiftInput {
  userId: string;
  locationId: string;
  startTime: string;
  endTime?: string;
  openingCash?: string;
  notes?: string;
}

export interface ShiftUpdate {
  endTime?: string;
  status?: string;
  openingCash?: string;
  closingCash?: string;
  notes?: string;
}

export interface Shift {
  id: string;
  userId: string;
  /** @nullable */
  userName?: string | null;
  locationId: string;
  /** @nullable */
  locationName?: string | null;
  startTime: string;
  /** @nullable */
  endTime?: string | null;
  status: string;
  /** @nullable */
  openingCash?: string | null;
  /** @nullable */
  closingCash?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt?: string;
}

export interface TransferInput {
  itemId: string;
  fromLocationId: string;
  toLocationId: string;
  quantity: number;
  notes?: string;
}

export interface Transfer {
  id: string;
  itemId: string;
  /** @nullable */
  itemName?: string | null;
  fromLocationId: string;
  /** @nullable */
  fromLocationName?: string | null;
  toLocationId: string;
  /** @nullable */
  toLocationName?: string | null;
  quantity: number;
  status: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  requestedById?: string | null;
  /** @nullable */
  approvedById?: string | null;
  createdAt?: string;
}

export interface AnalyticsSummary {
  totalSales: number;
  totalRevenue: string;
  totalTransactions: number;
  averageOrderValue: string;
  cashSales?: string;
  momoSales?: string;
  cardSales?: string;
  lowStockItems?: number;
}

export interface SalesByDay {
  date: string;
  revenue: string;
  transactions: number;
}

export interface TopItem {
  itemId: string;
  name: string;
  quantitySold: number;
  revenue: string;
}

export type MoMoInputNetwork = typeof MoMoInputNetwork[keyof typeof MoMoInputNetwork];


export const MoMoInputNetwork = {
  MTN: 'MTN',
  Telecel: 'Telecel',
} as const;

export interface MoMoInput {
  phone: string;
  network: MoMoInputNetwork;
  amount: string;
  reference: string;
  description?: string;
}

export interface MoMoResponse {
  success: boolean;
  reference: string;
  status: string;
  message?: string;
}

export interface MoMoStatus {
  reference: string;
  status: string;
  /** @nullable */
  amount?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  message?: string | null;
}

export type ListInventoryParams = {
locationId?: string;
categoryId?: string;
search?: string;
lowStock?: boolean;
};

export type ListTransactionsParams = {
locationId?: string;
startDate?: string;
endDate?: string;
paymentMethod?: string;
limit?: number;
offset?: number;
};

export type ListShiftsParams = {
locationId?: string;
userId?: string;
status?: string;
startDate?: string;
endDate?: string;
};

export type ListTransfersParams = {
locationId?: string;
status?: string;
};

export type GetAnalyticsSummaryParams = {
locationId?: string;
period?: GetAnalyticsSummaryPeriod;
};

export type GetAnalyticsSummaryPeriod = typeof GetAnalyticsSummaryPeriod[keyof typeof GetAnalyticsSummaryPeriod];


export const GetAnalyticsSummaryPeriod = {
  today: 'today',
  week: 'week',
  month: 'month',
  year: 'year',
} as const;

export type GetSalesByDayParams = {
locationId?: string;
days?: number;
};

export type GetTopItemsParams = {
locationId?: string;
limit?: number;
};