# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MirrorTech POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshInput"
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory items
      security:
        - bearerAuth: []
      parameters:
        - name: locationId
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create an inventory item
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get a single inventory item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update an inventory item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete an inventory item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
...
  /analytics/summary:
    get:
      operationId: getAnalyticsSummary
      tags: [analytics]
      summary: Dashboard summary metrics
      security:
        - bearerAuth: []
      parameters:
        - name: locationId
          in: query
          schema:
            type: string
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month, year]
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"
  /analytics/sales-by-day:
    get:
      operationId: getSalesByDay
      tags: [analytics]
      summary: Sales breakdown by day
      security:
        - bearerAuth: []
      parameters:
        - name: locationId
          in: query
          schema:
            type: string
        - name: days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Sales by day
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesByDay"
  /analytics/top-items:
    get:
      operationId: getTopItems
      tags: [analytics]
      summary: Top selling items
      security:
        - bearerAuth: []
      parameters:
        - name: locationId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopItem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    RefreshInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      required: [token, refreshToken, user]
      properties:
        token:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
    UserProfile:
      type: object
      required: [id, username, email, role]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          type: string
        locationId:
          type: ["string", "null"]
...
    Category:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    InventoryItem:
      type: object
      required: [id, name, price, quantity]
      properties:
        id:
          type: string
        name:
          type: string
        sku:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        price:
          type: string
        cost:
          type: ["string", "null"]
        quantity:
          type: integer
        minQuantity:
          type: integer
        locationId:
          type: ["string", "null"]
        categoryId:
          type: ["string", "null"]
        categoryName:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    InventoryInput:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
        sku:
          type: string
        description:
          type: string
        price:
          type: string
        cost:
          type: string
        quantity:
          type: integer
        minQuantity:
          type: integer
        locationId:
          type: string
        categoryId:
          type: string
        unit:
          type: string
    InventoryUpdate:
      type: object
      properties:
        name:
          type: string
        sku:
          type: string
        description:
          type: string
        price:
          type: string
        cost:
          type: string
        quantity:
          type: integer
        minQuantity:
          type: integer
        locationId:
          type: string
        categoryId:
          type: string
        unit:
          type: string
        isActive:
          type: boolean
...
    AnalyticsSummary:
      type: object
      required: [totalSales, totalRevenue, totalTransactions, averageOrderValue]
      properties:
        totalSales:
          type: number
        totalRevenue:
          type: string
        totalTransactions:
          type: integer
        averageOrderValue:
          type: string
        cashSales:
          type: string
        momoSales:
          type: string
        cardSales:
          type: string
        lowStockItems:
          type: integer
    SalesByDay:
      type: object
      required: [date, revenue, transactions]
      properties:
        date:
          type: string
        revenue:
          type: string
        transactions:
          type: integer
    TopItem:
      type: object
      required: [itemId, name, quantitySold, revenue]
      properties:
        itemId:
          type: string
        name:
          type: string
        quantitySold:
          type: integer
        revenue:
          type: string
...